+++ package.json
@@ -27,6 +27,7 @@
     "bootstrap": "^4.4.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-hook-form": "^5.5.1",
     "react-map-gl": "^5.2.3",
     "react-router-dom": "^5.1.2",
     "react-transition-group": "^4.3.0"
