+++ src/components/home/showcase/showcase.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import Boiler from './views/boiler';
 import Crusher from './views/crusher';
 import ShowcaseItem from './showcase-item/showcase-item';
-import Button from '../../shared/button/button';
+import Button from '../../shared/buttons/button';
 import * as buttonClassName from '../../../consts/button-types';
 import './showcase.scss';
