+++ src/components/shared/modal/modal.tsx
@@ -27,5 +27,4 @@ export default ({
       </div>
     </div>
   </CSSTransition>
-
 );
