+++ src/components/shared/order-form/order-form.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable jsx-a11y/label-has-associated-control */
 import React from 'react';
+import { useForm } from 'react-hook-form';
 import './order-form.scss';
-import Button from '../button/button';
+import ButtonSubmit from '../buttons/button-submit';
 import * as buttonTypes from '../../../consts/button-types';
 
 export default () => {
@@ -8,32 +10,32 @@ export default () => {
     category: 'Котел',
     model: 'Fit 20-784',
   };
+  const { register, handleSubmit, errors } = useForm();
+  const onSubmit = (data: object) => console.log(data);
   
   return (
     <div className="order-form" tabIndex={-1} role="dialog">
-      <form action="" method="" name="orderForm">
-        <fieldset>
-          <legend>Отправить заявку</legend>
-          <h3>Вы выбрали:</h3>
-          <p>
-            <span>
-              {myObj.category}
-            </span>
-            <span>
-              {myObj.model}
-            </span>
-          </p>
-          <p>
-            Мы получим Вашу заявку на эту модель и
-            <br />
-            перезвоним Вам, чтобы оформить заказ.
-          </p>
-          <div>
-            <label htmlFor="tel">Ваш номер телефона:</label>
-            <input type="tel" id="tel" name="tel" required placeholder="+7-999-000-00-00" />
-            <Button label="Отправить заявку" className={buttonTypes.longCallmeButton} clickHandler="" />
-          </div>
-        </fieldset>
+      <form onSubmit={handleSubmit(onSubmit)}>
+        <h3>Вы выбрали:</h3>
+        <p className="model">
+          <span>
+            {myObj.category}
+          </span>
+          <span>
+            {myObj.model}
+          </span>
+        </p>
+        <p className="comment">
+          Мы получим Вашу заявку на эту модель и
+          <br />
+          перезвоним Вам, чтобы оформить заказ.
+        </p>
+        <div>
+          <label htmlFor="tel">Ваш номер телефона:</label>
+          <input type="tel" id="tel" name="tel" ref={register({ required: true })} placeholder="+7-999-000-00-00" />
+          {errors.tel && <span className="error">Введите номер телефона</span>}
+          <ButtonSubmit label="Отправить заявку" className={buttonTypes.longCallmeButton} />
+        </div>
       </form>
     </div>
   );
